requirements.txt

streamlit
pandas
scikit-learn
requests
justwatch-python
